requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
imapclient>=2.3
celery>=5.3
redis>=5.0
cryptography>=41.0
python-dateutil>=2.8
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.6
dj-database-url>=2.1
